feat: VoucherExecution 함수형 인터페이스 설정
1. 기존 스프링 빈으로 사용되지 않는 VoucherExecution 구현 객체를 제거하고, 익명 클래스 사용
