+++ src/main/java/com/programmers/application/controller/command/execution/VoucherExecution.java
@@ -2,6 +2,7 @@ package com.programmers.application.controller.command.execution;
 
 import java.io.IOException;
 
+@FunctionalInterface
 public interface VoucherExecution {
-    void run() throws IOException;
+    void execute() throws IOException;
 }
